refactor : common - useCategory hook component
기존 useCategory 사용 페이지에서 useRef가 아닌 useState를 사용함에 따라 임시 리팩토링함

Related to : #99
